SOLR-14381: Grouping to use long instead of int for matches and ngroups to avoid overflows

+++ solr/core/src/java/org/apache/solr/request/SimpleFacets.java
@@ -335,7 +335,7 @@ public class SimpleFacets {
 
     AllGroupsCollector collector = new AllGroupsCollector<>(new TermGroupSelector(groupField));
     searcher.search(QueryUtils.combineQueryAndFilter(facetQuery, docSet.getTopFilter()), collector);
-    return collector.getGroupCount();
+    return (int)collector.getGroupCount();
   }
 
   enum FacetMethod {
